backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "profsim-backend"
version = "0.1.0"
description = "AI-driven Interactive Profession Simulation Platform"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",

    # Database
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",

    # Redis
    "redis[hiredis]==5.2.1",

    # Auth
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",

    # Validation & config
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "python-multipart==0.0.19",

    # AI
    "groq==0.13.0",
    "httpx==0.28.1",
    "rank-bm25==0.2.2",       # BM25 RAG retrieval — no external API needed
    "sse-starlette==2.1.3",   # Server-Sent Events for streaming feedback
    "nltk==3.9.1",            # Tokenisation for BM25

    # Utilities
    "python-dotenv==1.0.1",
    "orjson==3.10.12",
    "structlog==24.4.0",
    "tenacity==9.0.0",
    "anyio==4.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",
    "factory-boy==3.3.1",
    "faker==33.1.0",
    "ruff==0.8.4",
    "mypy==1.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
